kafka_producer/requirements.txt
kafka-python
psutil
speedtest-cli
numpy
